BE: few changes in the main and error files for standardization
