+++ server/app/shared/utils/logging.py
@@ -4,9 +4,32 @@ import logging
 import json
 from datetime import datetime
 from typing import Any, Dict, Optional
+import threading
 
 from fastapi import Request
 import uuid
+
+
+class RequestIdFilter(logging.Filter):
+    """Filter that adds request_id to log records."""
+    
+    _thread_local = threading.local()
+    
+    @classmethod
+    def set_request_id(cls, request_id: str):
+        """Set the request ID for the current thread."""
+        cls._thread_local.request_id = request_id
+    
+    @classmethod
+    def get_request_id(cls):
+        """Get the request ID for the current thread."""
+        return getattr(cls._thread_local, "request_id", "no-request-id")
+    
+    def filter(self, record):
+        """Add request_id to the log record."""
+        if not hasattr(record, "request_id"):
+            record.request_id = self.get_request_id()
+        return True
 
 
 class StructuredLogger:
@@ -75,6 +98,10 @@ async def request_id_middleware(request: Request, call_next):
     """Middleware to add a request ID to each request."""
     request_id = str(uuid.uuid4())
     request.state.request_id = request_id
+    
+    # Set the request ID for the current thread
+    RequestIdFilter.set_request_id(request_id)
+    
     response = await call_next(request)
     response.headers["X-Request-ID"] = request_id
     return response
